avances

+++ src/Routes/preguntas.js
@@ -8,12 +8,12 @@ const config = 'S~J?xm,:c7WU8HFz)K$a$N&[V:ez*EN#';
 
 function conectar(){
     const mysqlConnection = mysql.createConnection({
-        host: 'localhost',
-        user: 'root',
-        password: '03042021',
-        database: 'bdquetzual',
-        charset:  "utf-8",
-        multipleStatements: true
+        host: 'us-cdbr-east-05.cleardb.net',
+        user: 'b5281614b1195e',
+        password: '1025be05',
+        database: 'heroku_3f56efd65cc9251',
+        multipleStatements: true,
+        charset: 'UTF8_GENERAL_CI'
         });
     mysqlConnection.connect(function (err) {
         if (err) {
@@ -31,27 +31,31 @@ router.post('/Realizar/Pregunta', (req, res)=>{
     if(!token){
         res.json({'status': '¡ERROR!'});
     }else{
+        var seguirrr = false
         try{
             const tokend = jwt.verify(token, config);
+            seguirrr = true
         }catch(error){
             res.json({'status': '¡ERROR!'});
         }
-        const tokend = jwt.verify(token, config);
-        if(clave == claveusu && tokend.clave == claveusu && tokend.id == usu && tokend.id_rol == 1){
-            var mysqlConnection = conectar();
-            const query = 'insert into mpregunta (des_pre, fecha_pre, id_usup, id_estado) VALUES (?, ?, ?, ?)';
-            mysqlConnection.query(query, [pregunta, fecha, usu, estado], (err, rows)=>{
-                if(!err){
-                    mysqlConnection.destroy();
-                    res.json({'status': 'Guardado'});
-                }else{
-                    mysqlConnection.destroy();
-                    console.error(err);
-                    res.json({'status': '¡ERROR!'});
-                }
-            });
-        }else{
-            res.json({'status': '¡ERROR!'})
+        if(seguirrr){
+            const tokend = jwt.verify(token, config);
+            if(clave == claveusu && tokend.clave == claveusu && tokend.id == usu && tokend.id_rol == 1){
+                var mysqlConnection = conectar();
+                const query = 'insert into mpregunta (des_pre, fecha_pre, id_usup, id_estado) VALUES (?, ?, ?, ?)';
+                mysqlConnection.query(query, [pregunta, fecha, usu, estado], (err, rows)=>{
+                    if(!err){
+                        mysqlConnection.destroy();
+                        res.json({'status': 'Guardado'});
+                    }else{
+                        mysqlConnection.destroy();
+                        console.error(err);
+                        res.json({'status': '¡ERROR!'});
+                    }
+                });
+            }else{
+                res.json({'status': '¡ERROR!'})
+            }
         }
     }
 });
@@ -63,27 +67,31 @@ router.post('/Modificar/Pregunta/Pendiente', (req, res) =>{
     if(!token){
         res.json({'status': '¡ERROR!'});
     }else{
+        var seguirrr = false
         try{
             const tokend = jwt.verify(token, config);
+            seguirrr = true
         }catch(error){
             res.json({'status': '¡ERROR!'});
         }
-        const tokend = jwt.verify(token, config);
-        if(clave == claveusu && tokend.id_rol == 1 && tokend.id == usu && tokend.clave == claveusu){
-            var mysqlConnection = conectar();
-            const query = 'update mpregunta set des_pre = ? , fecha_pre = ? where id_pre = ? AND id_usup = ? AND id_estado = ?';
-            mysqlConnection.query(query, [pregunta, fecha, id_pre, usu, 1], (err, rows) =>{
-                if(!err){
-                    mysqlConnection.destroy();
-                    res.json({'status': 'Actualizado'})
-                }else{
-                    mysqlConnection.destroy();
-                    console.error(err);
-                    res.json({'status': '¡ERROR!'});
-                }
-            } );
-        }else{
-            res.json({'status': '¡ERROR!'});
+        if(seguirrr){
+            const tokend = jwt.verify(token, config);
+            if(clave == claveusu && tokend.id_rol == 1 && tokend.id == usu && tokend.clave == claveusu){
+                var mysqlConnection = conectar();
+                const query = 'update mpregunta set des_pre = ? , fecha_pre = ? where id_pre = ? AND id_usup = ? AND id_estado = ?';
+                mysqlConnection.query(query, [pregunta, fecha, id_pre, usu, 1], (err, rows) =>{
+                    if(!err){
+                        mysqlConnection.destroy();
+                        res.json({'status': 'Actualizado'})
+                    }else{
+                        mysqlConnection.destroy();
+                        console.error(err);
+                        res.json({'status': '¡ERROR!'});
+                    }
+                } );
+            }else{
+                res.json({'status': '¡ERROR!'});
+            }
         }
     }
 });
@@ -95,27 +103,31 @@ router.post('/Eliminar/Pregunta', (req, res)=>{
     if(!token){
         res.json({'status': '¡ERROR!'});
     }else{
+        var seguirrr = false
         try{
             const tokend = jwt.verify(token, config);
+            seguirrr = true
         }catch(error){
             res.json({'status': '¡ERROR!'});
         }
-        const tokend = jwt.verify(token, config);
-        if(clave == claveusu && tokend.id_rol == 1 && tokend.id == id_usu && tokend.clave == claveusu){
-            var mysqlConnection = conectar();
-            const query = 'delete from MPregunta where id_pre = ? AND id_usup = ? AND id_Estado = ?';
-            mysqlConnection.query(query,[pregunta, id_usu, 1], (err, rows)=>{
-                if(!err){
-                    mysqlConnection.destroy();
-                    res.json({'status': 'Eliminada'});
-                }else{
-                    mysqlConnection.destroy();
-                    console.error(err);
-                    res.json({'status': '¡ERROR!'});
-                }
-            });
-        }else{
-            res.json({'status': '¡ERROR!'})
+        if(seguirrr){
+            const tokend = jwt.verify(token, config);
+            if(clave == claveusu && tokend.id_rol == 1 && tokend.id == id_usu && tokend.clave == claveusu){
+                var mysqlConnection = conectar();
+                const query = 'delete from MPregunta where id_pre = ? AND id_usup = ? AND id_Estado = ?';
+                mysqlConnection.query(query,[pregunta, id_usu, 1], (err, rows)=>{
+                    if(!err){
+                        mysqlConnection.destroy();
+                        res.json({'status': 'Eliminada'});
+                    }else{
+                        mysqlConnection.destroy();
+                        console.error(err);
+                        res.json({'status': '¡ERROR!'});
+                    }
+                });
+            }else{
+                res.json({'status': '¡ERROR!'})
+            }
         }
     }
 });
@@ -127,48 +139,52 @@ router.post('/Usuario/Preguntas', (req, res)=>{
     if(!token){
         res.json({'status': '¡ERROR!'});
     }else{
+        var seguirrr = false
         try{
             const tokend = jwt.verify(token, config);
+            seguirrr = true
         }catch(error){
             res.json({'status': '¡ERROR!'});
         }
-        const tokend = jwt.verify(token, config);
-        if(clave == claveusu && id_usu == tokend.id && 1 == tokend.id_rol){
-           
-            var mysqlConnection = conectar();
-            if(estado == 3){
-                const query = 'select mpregunta.des_pre , mpregunta.fecha_pre , mrespuesta.des_res, mrespuesta.fecha_res  from mpregunta INNER JOIN mrespuesta ON mpregunta.id_pre = mrespuesta.id_pre where mpregunta.id_usup = ? AND mpregunta.id_estado = ? ';
-                mysqlConnection.query(query, [id_usu, estado], (err, rows) =>{
-                    if(!err){
-                        mysqlConnection.destroy();
-                        res.json({
-                            'status': 'Encontradas',
-                            'datos': rows
-                        });
-                    }else{
-                        mysqlConnection.destroy();
-                        console.error(err);
-                        res.json({'status': '¡ERROR!'});
-                    }
-                });
-            }else if(estado == 2 || estado == 1){
-                const query = 'select * from mpregunta where id_usup = ? AND id_estado = ?';
-                mysqlConnection.query(query, [id_usu, estado], (err, rows) =>{
-                    if(!err){
-                        mysqlConnection.destroy();
-                        res.json({
-                            'status': 'Encontradas',
-                            'datos': rows
-                        })
-                    }else{
-                        mysqlConnection.destroy();
-                        console.error(err);
-                        res.json({'status': '¡ERROR!'});
-                    }
-                });
+        if(seguirrr){
+            const tokend = jwt.verify(token, config);
+            if(clave == claveusu && id_usu == tokend.id && 1 == tokend.id_rol){
+            
+                var mysqlConnection = conectar();
+                if(estado == 3){
+                    const query = 'select mpregunta.des_pre , mpregunta.fecha_pre , mrespuesta.des_res, mrespuesta.fecha_res  from mpregunta INNER JOIN mrespuesta ON mpregunta.id_pre = mrespuesta.id_pre where mpregunta.id_usup = ? AND mpregunta.id_estado = ? ';
+                    mysqlConnection.query(query, [id_usu, estado], (err, rows) =>{
+                        if(!err){
+                            mysqlConnection.destroy();
+                            res.json({
+                                'status': 'Encontradas',
+                                'datos': rows
+                            });
+                        }else{
+                            mysqlConnection.destroy();
+                            console.error(err);
+                            res.json({'status': '¡ERROR!'});
+                        }
+                    });
+                }else if(estado == 2 || estado == 1){
+                    const query = 'select * from mpregunta where id_usup = ? AND id_estado = ?';
+                    mysqlConnection.query(query, [id_usu, estado], (err, rows) =>{
+                        if(!err){
+                            mysqlConnection.destroy();
+                            res.json({
+                                'status': 'Encontradas',
+                                'datos': rows
+                            })
+                        }else{
+                            mysqlConnection.destroy();
+                            console.error(err);
+                            res.json({'status': '¡ERROR!'});
+                        }
+                    });
+                }
+            }else{
+                res.json({'status': '¡ERROR!'})
             }
-        }else{
-            res.json({'status': '¡ERROR!'})
         }
     }
 });
@@ -178,7 +194,7 @@ router.post('/Respondidas/Actuales', (req, res)=>{
     const {Clave, clasi} = req.body;
     if(Clave == claveusu || Clave == clavedoc){
         var mysqlConnection = conectar();
-        if(clasi == 0 || clasi == 6){
+        if(clasi == 0 || clasi >= 6){
             const query = 'select mpregunta.id_pre, mpregunta.des_pre, mpregunta.fecha_pre, musuario.fecha_nac from mpregunta INNER JOIN eusuario ON mpregunta.id_usup = eusuario.id_enusuario INNER JOIN musuario ON eusuario.id_usu = musuario.id_usu where mpregunta.id_estado = ? ORDER BY mpregunta.id_pre DESC';
             mysqlConnection.query(query, 2, (err, rows)=>{
                 if(!err){
@@ -221,30 +237,34 @@ router.post('/Pendientes', (req, res)=>{
     if(!token){
         res.json({'status': '¡ERROR!'});
     }else{
+        var seguirrr = false
         try{
             const tokend = jwt.verify(token, config);
+            seguirrr = true
         }catch(error){
             res.json({'status': '¡ERROR!'});
         }
-        const tokend = jwt.verify(token, config);
-        if(clave == clavedoc && tokend.id_rol == 2 && tokend.clave == clavedoc){
-            var mysqlConnection = conectar();
-            const query = 'select * from mpregunta where id_estado = 1';
-            mysqlConnection.query(query, (err, rows)=>{
-                if(!err){
-                    mysqlConnection.destroy();
-                        res.json({
-                            'status': 'Encontradas',
-                            'datos': rows
-                    });
-                }else{
-                    mysqlConnection.destroy();
-                    console.error(err);
-                    res.json({'status': '¡ERROR!'});
-                }
-            });
-        }else{
-            res.json({'status': '¡ERROR!'})
+        if(seguirrr){
+            const tokend = jwt.verify(token, config);
+            if(clave == clavedoc && tokend.id_rol == 2 && tokend.clave == clavedoc){
+                var mysqlConnection = conectar();
+                const query = 'select * from mpregunta where id_estado = 1';
+                mysqlConnection.query(query, (err, rows)=>{
+                    if(!err){
+                        mysqlConnection.destroy();
+                            res.json({
+                                'status': 'Encontradas',
+                                'datos': rows
+                        });
+                    }else{
+                        mysqlConnection.destroy();
+                        console.error(err);
+                        res.json({'status': '¡ERROR!'});
+                    }
+                });
+            }else{
+                res.json({'status': '¡ERROR!'})
+            }
         }
     }
 });
@@ -317,5 +337,8 @@ router.post('/Similares', (req, res)=>{
         res.json({'status': '¡ERROR!'})
     }
 });
+
+
+
 
 module.exports = router;
